test(di): make parameter extraction tests feature-safe
